scripts/requirements.txt
# =============================================================================
# RAG Document Assistant - Script Dependencies
# =============================================================================
# Includes document processing libraries for data ingestion

# Environment Management
python-dotenv>=1.0.1

# LangChain and Azure OpenAI
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.8

# Vector Database
qdrant-client[fastembed]>=1.14.0

# Document Loaders
pypdf>=4.0.0
python-docx>=1.1.0
unstructured>=0.14.0
openpyxl>=3.1.0
pdfminer.six>=20231228
docx2txt>=0.8

# Image Processing
pillow>=9.0.0
pillow-heif>=0.16.0
matplotlib>=3.5.0
pytesseract>=0.3.10

# Utilities
tqdm>=4.65.0
tenacity>=8.2.0
